Function libduckdb_sys::duckdb_add_replacement_scan

pub unsafe extern "C" fn duckdb_add_replacement_scan(
    db: duckdb_database,
    replacement: duckdb_replacement_callback_t,
    extra_data: *mut c_void,
    delete_callback: duckdb_delete_callback_t,
)

Add a replacement scan definition to the specified database.

db: The database object to add the replacement scan to replacement: The replacement scan callback extra_data: Extra data that is passed back into the specified callback delete_callback: The delete callback to call on the extra data, if any