Function libduckdb_sys::duckdb_arrow_array_scan

pub unsafe extern "C" fn duckdb_arrow_array_scan(
    connection: duckdb_connection,
    table_name: *const c_char,
    arrow_schema: duckdb_arrow_schema,
    arrow_array: duckdb_arrow_array,
    out_stream: *mut duckdb_arrow_stream,
) -> duckdb_state

Scans the Arrow array and creates a view with the given name. Note that after running duckdb_arrow_array_scan, duckdb_destroy_arrow_stream must be called on the out stream.

connection: The connection on which to execute the scan. table_name: Name of the temporary view to create. arrow_schema: Arrow schema wrapper. arrow_array: Arrow array wrapper. out_stream: Output array stream that wraps around the passed schema, for releasing/deleting once done. returns: DuckDBSuccess on success or DuckDBError on failure.