Function libduckdb_sys::duckdb_get_config_flag

pub unsafe extern "C" fn duckdb_get_config_flag(
    index: usize,
    out_name: *mut *const c_char,
    out_description: *mut *const c_char,
) -> duckdb_state

Obtains a human-readable name and description of a specific configuration option. This can be used to e.g. display configuration options. This will succeed unless index is out of range (i.e. >= duckdb_config_count).

The result name or description MUST NOT be freed.

index: The index of the configuration option (between 0 and duckdb_config_count) out_name: A name of the configuration flag. out_description: A description of the configuration flag. returns: DuckDBSuccess on success or DuckDBError on failure.