Function libduckdb_sys::duckdb_nullmask_data

pub unsafe extern "C" fn duckdb_nullmask_data(
    result: *mut duckdb_result,
    col: idx_t,
) -> *mut bool

DEPRECATED**: Prefer using duckdb_result_get_chunk instead.

Returns the nullmask of a specific column of a result in columnar format. The nullmask indicates for every row whether or not the corresponding row is NULL. If a row is NULL, the values present in the array provided by duckdb_column_data are undefined.

int32_t *data = (int32_t *) duckdb_column_data(&result, 0);
bool *nullmask = duckdb_nullmask_data(&result, 0);
if (nullmask[row]) {
printf("Data for row %d: NULL\n", row);
} else {
printf("Data for row %d: %d\n", row, data[row]);
}

result: The result object to fetch the nullmask from. col: The column index. returns: The nullmask of the specified column.