Function libduckdb_sys::duckdb_query_arrow

pub unsafe extern "C" fn duckdb_query_arrow(
    connection: duckdb_connection,
    query: *const c_char,
    out_result: *mut duckdb_arrow,
) -> duckdb_state

Executes a SQL query within a connection and stores the full (materialized) result in an arrow structure. If the query fails to execute, DuckDBError is returned and the error message can be retrieved by calling duckdb_query_arrow_error.

Note that after running duckdb_query_arrow, duckdb_destroy_arrow must be called on the result object even if the query fails, otherwise the error stored within the result will not be freed correctly.

connection: The connection to perform the query in. query: The SQL query to run. out_result: The query result. returns: DuckDBSuccess on success or DuckDBError on failure.