Crate libduckdb_sys

Structs

Error

__fsid_t

_duckdb_appender

! The appender enables fast data loading into DuckDB. ! Must be destroyed with duckdb_appender_destroy.

_duckdb_arrow

! Holds an arrow query result. Must be destroyed with duckdb_destroy_arrow.

_duckdb_arrow_array

! Holds an arrow array. Remember to release the respective ArrowArray object.

_duckdb_arrow_schema

! Holds an arrow schema. Remember to release the respective ArrowSchema object.

_duckdb_arrow_stream

! Holds an arrow array stream. Must be destroyed with duckdb_destroy_arrow_stream.

_duckdb_config

! Can be used to provide start-up options for the DuckDB instance. ! Must be destroyed with duckdb_destroy_config.

_duckdb_connection

! A connection to a duckdb database. Must be closed with duckdb_disconnect.

_duckdb_data_chunk

! Contains a data chunk from a duckdb_result. ! Must be destroyed with duckdb_destroy_data_chunk.

_duckdb_database

! A database object. Should be closed with duckdb_close.

_duckdb_extracted_statements

! Extracted statements. Must be destroyed with duckdb_destroy_extracted.

_duckdb_logical_type

! Holds an internal logical type. ! Must be destroyed with duckdb_destroy_logical_type.

_duckdb_pending_result

! The pending result represents an intermediate structure for a query that is not yet fully executed. ! Must be destroyed with duckdb_destroy_pending.

_duckdb_prepared_statement

! A prepared statement is a parameterized query that allows you to bind parameters to it. ! Must be destroyed with duckdb_destroy_prepare.

_duckdb_value

! Holds a DuckDB value, which wraps a type. ! Must be destroyed with duckdb_destroy_value.

_duckdb_vector

! A vector to a specified column in a data chunk. Lives as long as the ! data chunk lives, i.e., must not be destroyed.

duckdb_blob

! BLOBs are composed of a byte pointer and a size. You must free blob.data ! with duckdb_free.

duckdb_column

! A column consists of a pointer to its internal data. Don’t operate on this type directly. ! Instead, use functions such as duckdb_column_data, duckdb_nullmask_data, ! duckdb_column_type, and duckdb_column_name, which take the result and the column index ! as their parameters

duckdb_date

! Days are stored as days since 1970-01-01 ! Use the duckdb_from_date/duckdb_to_date function to extract individual information

duckdb_date_struct

duckdb_decimal

! Decimals are composed of a width and a scale, and are stored in a hugeint

duckdb_hugeint

! Hugeints are composed of a (lower, upper) component ! The value of the hugeint is upper * 2^64 + lower ! For easy usage, the functions duckdb_hugeint_to_double/duckdb_double_to_hugeint are recommended

duckdb_interval

duckdb_list_entry

! The internal representation of a list metadata entry contains the list’s offset in ! the child vector, and its length. The parent vector holds these metadata entries, ! whereas the child vector holds the data

duckdb_query_progress_type

! A type holding information about the query execution progress

duckdb_result

! A query result consists of a pointer to its internal data. ! Must be freed with ‘duckdb_destroy_result’.

duckdb_string

! Strings are composed of a char pointer and a size. You must free string.data ! with duckdb_free.

duckdb_string_t

! The internal representation of a VARCHAR (string_t). If the VARCHAR does not ! exceed 12 characters, then we inline it. Otherwise, we inline a prefix for faster ! string comparisons and store a pointer to the remaining characters. This is a non- ! owning structure, i.e., it does not have to be freed.

duckdb_string_t__bindgen_ty_1__bindgen_ty_1

duckdb_string_t__bindgen_ty_1__bindgen_ty_2

duckdb_time

! Time is stored as microseconds since 00:00:00 ! Use the duckdb_from_time/duckdb_to_time function to extract individual information

duckdb_time_struct

duckdb_time_tz

! TIME_TZ is stored as 40 bits for int64_t micros, and 24 bits for int32_t offset

duckdb_time_tz_struct

duckdb_timestamp

! Timestamps are stored as microseconds since 1970-01-01 ! Use the duckdb_from_timestamp/duckdb_to_timestamp function to extract individual information

duckdb_timestamp_struct

duckdb_uhugeint

max_align_t

Enums

ErrorCode

Error Codes

Constants

DUCKDB_API_0_3_1

DUCKDB_API_0_3_2

DUCKDB_API_LATEST

DUCKDB_API_VERSION

DUCKDB_TYPE_DUCKDB_TYPE_ARRAY

DUCKDB_TYPE_DUCKDB_TYPE_BIGINT

DUCKDB_TYPE_DUCKDB_TYPE_BIT

DUCKDB_TYPE_DUCKDB_TYPE_BLOB

DUCKDB_TYPE_DUCKDB_TYPE_BOOLEAN

DUCKDB_TYPE_DUCKDB_TYPE_DATE

DUCKDB_TYPE_DUCKDB_TYPE_DECIMAL

DUCKDB_TYPE_DUCKDB_TYPE_DOUBLE

DUCKDB_TYPE_DUCKDB_TYPE_ENUM

DUCKDB_TYPE_DUCKDB_TYPE_FLOAT

DUCKDB_TYPE_DUCKDB_TYPE_HUGEINT

DUCKDB_TYPE_DUCKDB_TYPE_INTEGER

DUCKDB_TYPE_DUCKDB_TYPE_INTERVAL

DUCKDB_TYPE_DUCKDB_TYPE_INVALID

DUCKDB_TYPE_DUCKDB_TYPE_LIST

DUCKDB_TYPE_DUCKDB_TYPE_MAP

DUCKDB_TYPE_DUCKDB_TYPE_SMALLINT

DUCKDB_TYPE_DUCKDB_TYPE_STRUCT

DUCKDB_TYPE_DUCKDB_TYPE_TIME

DUCKDB_TYPE_DUCKDB_TYPE_TIMESTAMP

DUCKDB_TYPE_DUCKDB_TYPE_TIMESTAMP_MS

DUCKDB_TYPE_DUCKDB_TYPE_TIMESTAMP_NS

DUCKDB_TYPE_DUCKDB_TYPE_TIMESTAMP_S

DUCKDB_TYPE_DUCKDB_TYPE_TIMESTAMP_TZ

DUCKDB_TYPE_DUCKDB_TYPE_TIME_TZ

DUCKDB_TYPE_DUCKDB_TYPE_TINYINT

DUCKDB_TYPE_DUCKDB_TYPE_UBIGINT

DUCKDB_TYPE_DUCKDB_TYPE_UHUGEINT

DUCKDB_TYPE_DUCKDB_TYPE_UINTEGER

DUCKDB_TYPE_DUCKDB_TYPE_UNION

DUCKDB_TYPE_DUCKDB_TYPE_USMALLINT

DUCKDB_TYPE_DUCKDB_TYPE_UTINYINT

DUCKDB_TYPE_DUCKDB_TYPE_UUID

DUCKDB_TYPE_DUCKDB_TYPE_VARCHAR

DuckDBError

DuckDBSuccess

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

PTRDIFF_MAX

PTRDIFF_MIN

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WINT_MAX

WINT_MIN

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_FEATURES_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_SYS_CDEFS_H

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_IEC_60559_BFP__

__STDC_IEC_60559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__bool_true_false_are_defined

__glibc_c99_flexarr_available

duckdb_pending_state_DUCKDB_PENDING_ERROR

duckdb_pending_state_DUCKDB_PENDING_NO_TASKS_AVAILABLE

duckdb_pending_state_DUCKDB_PENDING_RESULT_NOT_READY

duckdb_pending_state_DUCKDB_PENDING_RESULT_READY

duckdb_result_type_DUCKDB_RESULT_TYPE_CHANGED_ROWS

duckdb_result_type_DUCKDB_RESULT_TYPE_INVALID

duckdb_result_type_DUCKDB_RESULT_TYPE_NOTHING

duckdb_result_type_DUCKDB_RESULT_TYPE_QUERY_RESULT

duckdb_state_DuckDBError

duckdb_state_DuckDBSuccess

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_ALTER

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_ANALYZE

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_ATTACH

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_CALL

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_COPY

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_CREATE

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_CREATE_FUNC

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_DELETE

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_DETACH

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_DROP

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_EXECUTE

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_EXPLAIN

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_EXPORT

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_EXTENSION

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_INSERT

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_INVALID

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_LOAD

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_LOGICAL_PLAN

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_MULTI

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_PRAGMA

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_PREPARE

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_RELATION

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_SELECT

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_SET

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_TRANSACTION

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_UPDATE

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_VACUUM

duckdb_statement_type_DUCKDB_STATEMENT_TYPE_VARIABLE_SET

false_

true_

Functions

code_to_str

duckdb_add_replacement_scan⚠

Add a replacement scan definition to the specified database.

duckdb_append_blob⚠

Append a blob value to the appender.

duckdb_append_bool⚠

Append a bool value to the appender.

duckdb_append_data_chunk⚠

Appends a pre-filled data chunk to the specified appender.

duckdb_append_date⚠

Append a duckdb_date value to the appender.

duckdb_append_double⚠

Append a double value to the appender.

duckdb_append_float⚠

Append a float value to the appender.

duckdb_append_hugeint⚠

Append a duckdb_hugeint value to the appender.

duckdb_append_int8⚠

Append an int8_t value to the appender.

duckdb_append_int16⚠

Append an int16_t value to the appender.

duckdb_append_int32⚠

Append an int32_t value to the appender.

duckdb_append_int64⚠

Append an int64_t value to the appender.

duckdb_append_interval⚠

Append a duckdb_interval value to the appender.

duckdb_append_null⚠

Append a NULL value to the appender (of any type).

duckdb_append_time⚠

Append a duckdb_time value to the appender.

duckdb_append_timestamp⚠

Append a duckdb_timestamp value to the appender.

duckdb_append_uhugeint⚠

Append a duckdb_uhugeint value to the appender.

duckdb_append_uint8⚠

Append a uint8_t value to the appender.

duckdb_append_uint16⚠

Append a uint16_t value to the appender.

duckdb_append_uint32⚠

Append a uint32_t value to the appender.

duckdb_append_uint64⚠

Append a uint64_t value to the appender.

duckdb_append_varchar⚠

Append a varchar value to the appender.

duckdb_append_varchar_length⚠

Append a varchar value to the appender.

duckdb_appender_begin_row⚠

A nop function, provided for backwards compatibility reasons. Does nothing. Only duckdb_appender_end_row is required.

duckdb_appender_close⚠

Close the appender, flushing all intermediate state in the appender to the table and closing it for further appends.

duckdb_appender_column_count⚠

Returns the number of columns in the table that belongs to the appender.

duckdb_appender_column_type⚠

Returns the type of the column at the specified index.

duckdb_appender_create⚠

Creates an appender object.

duckdb_appender_destroy⚠

Close the appender and destroy it. Flushing all intermediate state in the appender to the table, and de-allocating all memory associated with the appender.

duckdb_appender_end_row⚠

Finish the current row of appends. After end_row is called, the next row can be appended.

duckdb_appender_error⚠

Returns the error message associated with the given appender. If the appender has no error message, this returns nullptr instead.

duckdb_appender_flush⚠

Flush the appender to the table, forcing the cache of the appender to be cleared and the data to be appended to the base table.

duckdb_array_type_array_size⚠

Retrieves the array size of the given array type.

duckdb_array_type_child_type⚠

Retrieves the child type of the given array type.

duckdb_array_vector_get_child⚠

Retrieves the child vector of a array vector.

duckdb_arrow_array_scan⚠

Scans the Arrow array and creates a view with the given name. Note that after running duckdb_arrow_array_scan, duckdb_destroy_arrow_stream must be called on the out stream.

duckdb_arrow_column_count⚠

Returns the number of columns present in the arrow result object.

duckdb_arrow_row_count⚠

Returns the number of rows present in the arrow result object.

duckdb_arrow_rows_changed⚠

Returns the number of rows changed by the query stored in the arrow result. This is relevant only for INSERT/UPDATE/DELETE queries. For other queries the rows_changed will be 0.

duckdb_arrow_scan⚠

Scans the Arrow stream and creates a view with the given name.

duckdb_bind_add_result_column⚠

Adds a result column to the output of the table function.

duckdb_bind_blob⚠

Binds a blob value to the prepared statement at the specified index.

duckdb_bind_boolean⚠

Binds a bool value to the prepared statement at the specified index.

duckdb_bind_date⚠

Binds a duckdb_date value to the prepared statement at the specified index.

duckdb_bind_decimal⚠

Binds a duckdb_decimal value to the prepared statement at the specified index.

duckdb_bind_double⚠

Binds a double value to the prepared statement at the specified index.

duckdb_bind_float⚠

Binds a float value to the prepared statement at the specified index.

duckdb_bind_get_extra_info⚠

Retrieves the extra info of the function as set in duckdb_table_function_set_extra_info.

duckdb_bind_get_named_parameter⚠

Retrieves a named parameter with the given name.

duckdb_bind_get_parameter⚠

Retrieves the parameter at the given index.

duckdb_bind_get_parameter_count⚠

Retrieves the number of regular (non-named) parameters to the function.

duckdb_bind_hugeint⚠

Binds a duckdb_hugeint value to the prepared statement at the specified index.

duckdb_bind_int8⚠

Binds an int8_t value to the prepared statement at the specified index.

duckdb_bind_int16⚠

Binds an int16_t value to the prepared statement at the specified index.

duckdb_bind_int32⚠

Binds an int32_t value to the prepared statement at the specified index.

duckdb_bind_int64⚠

Binds an int64_t value to the prepared statement at the specified index.

duckdb_bind_interval⚠

Binds a duckdb_interval value to the prepared statement at the specified index.

duckdb_bind_null⚠

Binds a NULL value to the prepared statement at the specified index.

duckdb_bind_parameter_index⚠

Retrieve the index of the parameter for the prepared statement, identified by name

duckdb_bind_set_bind_data⚠

Sets the user-provided bind data in the bind object. This object can be retrieved again during execution.

duckdb_bind_set_cardinality⚠

Sets the cardinality estimate for the table function, used for optimization.

duckdb_bind_set_error⚠

Report that an error has occurred while calling bind.

duckdb_bind_time⚠

Binds a duckdb_time value to the prepared statement at the specified index.

duckdb_bind_timestamp⚠

Binds a duckdb_timestamp value to the prepared statement at the specified index.

duckdb_bind_uhugeint⚠

Binds an duckdb_uhugeint value to the prepared statement at the specified index.

duckdb_bind_uint8⚠

Binds an uint8_t value to the prepared statement at the specified index.

duckdb_bind_uint16⚠

Binds an uint16_t value to the prepared statement at the specified index.

duckdb_bind_uint32⚠

Binds an uint32_t value to the prepared statement at the specified index.

duckdb_bind_uint64⚠

Binds an uint64_t value to the prepared statement at the specified index.

duckdb_bind_value⚠

Binds a value to the prepared statement at the specified index.

duckdb_bind_varchar⚠

Binds a null-terminated varchar value to the prepared statement at the specified index.

duckdb_bind_varchar_length⚠

Binds a varchar value to the prepared statement at the specified index.

duckdb_clear_bindings⚠

Clear the params bind to the prepared statement.

duckdb_close⚠

Closes the specified database and de-allocates all memory allocated for that database. This should be called after you are done with any database allocated through duckdb_open or duckdb_open_ext. Note that failing to call duckdb_close (in case of e.g. a program crash) will not cause data corruption. Still, it is recommended to always correctly close a database object after you are done with it.

duckdb_column_count⚠

Returns the number of columns present in a the result object.

duckdb_column_data⚠

DEPRECATED**: Prefer using duckdb_result_get_chunk instead.

duckdb_column_logical_type⚠

Returns the logical column type of the specified column.

duckdb_column_name⚠

Returns the column name of the specified column. The result should not need to be freed; the column names will automatically be destroyed when the result is destroyed.

duckdb_column_type⚠

Returns the column type of the specified column.

duckdb_config_count⚠

This returns the total amount of configuration options available for usage with duckdb_get_config_flag.

duckdb_connect⚠

Opens a connection to a database. Connections are required to query the database, and store transactional state associated with the connection. The instantiated connection should be closed using ‘duckdb_disconnect’.

duckdb_create_array_type⚠

Creates a array type from its child type. The resulting type should be destroyed with duckdb_destroy_logical_type.

duckdb_create_array_value⚠

Creates a array value from a type and an array of values of length value_count

duckdb_create_config⚠

Initializes an empty configuration object that can be used to provide start-up options for the DuckDB instance through duckdb_open_ext. The duckdb_config must be destroyed using ‘duckdb_destroy_config’

duckdb_create_data_chunk⚠

Creates an empty DataChunk with the specified set of types.

duckdb_create_decimal_type⚠

Creates a duckdb_logical_type of type decimal with the specified width and scale. The resulting type should be destroyed with duckdb_destroy_logical_type.

duckdb_create_enum_type⚠

Creates an ENUM type from the passed member name array. The resulting type should be destroyed with duckdb_destroy_logical_type.

duckdb_create_int64⚠

Creates a value from an int64

duckdb_create_list_type⚠

Creates a list type from its child type. The resulting type should be destroyed with duckdb_destroy_logical_type.

duckdb_create_list_value⚠

Creates a list value from a type and an array of values of length value_count

duckdb_create_logical_type⚠

Creates a duckdb_logical_type from a standard primitive type. The resulting type should be destroyed with duckdb_destroy_logical_type.

duckdb_create_map_type⚠

Creates a map type from its key type and value type. The resulting type should be destroyed with duckdb_destroy_logical_type.

duckdb_create_struct_type⚠

Creates a STRUCT type from the passed member name and type arrays. The resulting type should be destroyed with duckdb_destroy_logical_type.

duckdb_create_struct_value⚠

Creates a struct value from a type and an array of values

duckdb_create_table_function⚠

Creates a new empty table function.

duckdb_create_task_state⚠

Creates a task state that can be used with duckdb_execute_tasks_state to execute tasks until duckdb_finish_execution is called on the state.

duckdb_create_time_tz⚠

Create a duckdb_time_tz object from micros and a timezone offset.

duckdb_create_union_type⚠

Creates a UNION type from the passed types array. The resulting type should be destroyed with duckdb_destroy_logical_type.

duckdb_create_varchar⚠

Creates a value from a null-terminated string

duckdb_create_varchar_length⚠

Creates a value from a string

duckdb_data_chunk_get_column_count⚠

Retrieves the number of columns in a data chunk.

duckdb_data_chunk_get_size⚠

Retrieves the current number of tuples in a data chunk.

duckdb_data_chunk_get_vector⚠

Retrieves the vector at the specified column index in the data chunk.

duckdb_data_chunk_reset⚠

Resets a data chunk, clearing the validity masks and setting the cardinality of the data chunk to 0.

duckdb_data_chunk_set_size⚠

Sets the current number of tuples in a data chunk.

duckdb_decimal_internal_type⚠

Retrieves the internal storage type of a decimal type.

duckdb_decimal_scale⚠

Retrieves the scale of a decimal type.

duckdb_decimal_to_double⚠

Converts a duckdb_decimal object (as obtained from a DUCKDB_TYPE_DECIMAL column) into a double.

duckdb_decimal_width⚠

Retrieves the width of a decimal type.

duckdb_destroy_arrow⚠

Closes the result and de-allocates all memory allocated for the arrow result.

duckdb_destroy_arrow_stream⚠

Releases the arrow array stream and de-allocates its memory.

duckdb_destroy_config⚠

Destroys the specified configuration object and de-allocates all memory allocated for the object.

duckdb_destroy_data_chunk⚠

Destroys the data chunk and de-allocates all memory allocated for that chunk.

duckdb_destroy_extracted⚠

De-allocates all memory allocated for the extracted statements. extracted_statements: The extracted statements to destroy.

duckdb_destroy_logical_type⚠

Destroys the logical type and de-allocates all memory allocated for that type.

duckdb_destroy_pending⚠

Closes the pending result and de-allocates all memory allocated for the result.

duckdb_destroy_prepare⚠

Closes the prepared statement and de-allocates all memory allocated for the statement.

duckdb_destroy_result⚠

Closes the result and de-allocates all memory allocated for that connection.

duckdb_destroy_table_function⚠

Destroys the given table function object.

duckdb_destroy_task_state⚠

Destroys the task state returned from duckdb_create_task_state.

duckdb_destroy_value⚠

Destroys the value and de-allocates all memory allocated for that type.

duckdb_disconnect⚠

Closes the specified connection and de-allocates all memory allocated for that connection.

duckdb_double_to_decimal⚠

Converts a double value to a duckdb_decimal object.

duckdb_double_to_hugeint⚠

Converts a double value to a duckdb_hugeint object.

duckdb_double_to_uhugeint⚠

Converts a double value to a duckdb_uhugeint object.

duckdb_enum_dictionary_size⚠

Retrieves the dictionary size of the enum type.

duckdb_enum_dictionary_value⚠

Retrieves the dictionary value at the specified position from the enum.

duckdb_enum_internal_type⚠

Retrieves the internal storage type of an enum type.

duckdb_execute_n_tasks_state⚠

Execute DuckDB tasks on this thread.

duckdb_execute_pending⚠

Fully execute a pending query result, returning the final query result.

duckdb_execute_prepared⚠

Executes the prepared statement with the given bound parameters, and returns a materialized query result.

duckdb_execute_prepared_arrow⚠

Executes the prepared statement with the given bound parameters, and returns an arrow query result. Note that after running duckdb_execute_prepared_arrow, duckdb_destroy_arrow must be called on the result object.

duckdb_execute_prepared_streaming⚠

Executes the prepared statement with the given bound parameters, and returns an optionally-streaming query result. To determine if the resulting query was in fact streamed, use duckdb_result_is_streaming

duckdb_execute_tasks⚠

Execute DuckDB tasks on this thread.

duckdb_execute_tasks_state⚠

Execute DuckDB tasks on this thread.

duckdb_execution_is_finished⚠

Returns true if the execution of the current query is finished.

duckdb_extract_statements⚠

Extract all statements from a query. Note that after calling duckdb_extract_statements, the extracted statements should always be destroyed using duckdb_destroy_extracted, even if no statements were extracted.

duckdb_extract_statements_error⚠

Returns the error message contained within the extracted statements. The result of this function must not be freed. It will be cleaned up when duckdb_destroy_extracted is called.

duckdb_finish_execution⚠

Finish execution on a specific task.

duckdb_free⚠

Free a value returned from duckdb_malloc, duckdb_value_varchar, duckdb_value_blob, or duckdb_value_string.

duckdb_from_date⚠

Decompose a duckdb_date object into year, month and date (stored as duckdb_date_struct).

duckdb_from_time⚠

Decompose a duckdb_time object into hour, minute, second and microsecond (stored as duckdb_time_struct).

duckdb_from_time_tz⚠

Decompose a TIME_TZ objects into micros and a timezone offset.

duckdb_from_timestamp⚠

Decompose a duckdb_timestamp object into a duckdb_timestamp_struct.

duckdb_function_get_bind_data⚠

Gets the bind data set by duckdb_bind_set_bind_data during the bind.

duckdb_function_get_extra_info⚠

Retrieves the extra info of the function as set in duckdb_table_function_set_extra_info.

duckdb_function_get_init_data⚠

Gets the init data set by duckdb_init_set_init_data during the init.

duckdb_function_get_local_init_data⚠

Gets the thread-local init data set by duckdb_init_set_init_data during the local_init.

duckdb_function_set_error⚠

Report that an error has occurred while executing the function.

duckdb_get_config_flag⚠

Obtains a human-readable name and description of a specific configuration option. This can be used to e.g. display configuration options. This will succeed unless index is out of range (i.e. >= duckdb_config_count).

duckdb_get_int64⚠

Obtains an int64 of the given value.

duckdb_get_type_id⚠

Retrieves the enum type class of a duckdb_logical_type.

duckdb_get_varchar⚠

Obtains a string representation of the given value. The result must be destroyed with duckdb_free.

duckdb_hugeint_to_double⚠

Converts a duckdb_hugeint object (as obtained from a DUCKDB_TYPE_HUGEINT column) into a double.

duckdb_init_get_bind_data⚠

Gets the bind data set by duckdb_bind_set_bind_data during the bind.

duckdb_init_get_column_count⚠

Returns the number of projected columns.

duckdb_init_get_column_index⚠

Returns the column index of the projected column at the specified position.

duckdb_init_get_extra_info⚠

Retrieves the extra info of the function as set in duckdb_table_function_set_extra_info.

duckdb_init_set_error⚠

Report that an error has occurred while calling init.

duckdb_init_set_init_data⚠

Sets the user-provided init data in the init object. This object can be retrieved again during execution.

duckdb_init_set_max_threads⚠

Sets how many threads can process this table function in parallel (default: 1)

duckdb_interrupt⚠

Interrupt running query

duckdb_is_finite_date⚠

Test a duckdb_date to see if it is a finite value.

duckdb_is_finite_timestamp⚠

Test a duckdb_timestamp to see if it is a finite value.

duckdb_library_version⚠

Returns the version of the linked DuckDB, with a version postfix for dev versions

duckdb_list_type_child_type⚠

Retrieves the child type of the given list type.

duckdb_list_vector_get_child⚠

Retrieves the child vector of a list vector.

duckdb_list_vector_get_size⚠

Returns the size of the child vector of the list.

duckdb_list_vector_reserve⚠

Sets the total capacity of the underlying child-vector of a list.

duckdb_list_vector_set_size⚠

Sets the total size of the underlying child-vector of a list vector.

duckdb_logical_type_get_alias⚠

Returns the alias of a duckdb_logical_type, if one is set, else NULL. The result must be destroyed with duckdb_free.

duckdb_malloc⚠

Allocate size bytes of memory using the duckdb internal malloc function. Any memory allocated in this manner should be freed using duckdb_free.

duckdb_map_type_key_type⚠

Retrieves the key type of the given map type.

duckdb_map_type_value_type⚠

Retrieves the value type of the given map type.

duckdb_nparams⚠

Returns the number of parameters that can be provided to the given prepared statement.

duckdb_nullmask_data⚠

DEPRECATED**: Prefer using duckdb_result_get_chunk instead.

duckdb_open⚠

Creates a new database or opens an existing database file stored at the given path. If no path is given a new in-memory database is created instead. The instantiated database should be closed with ‘duckdb_close’.

duckdb_open_ext⚠

Extended version of duckdb_open. Creates a new database or opens an existing database file stored at the given path. The instantiated database should be closed with ‘duckdb_close’.

duckdb_param_type⚠

Returns the parameter type for the parameter at the given index.

duckdb_parameter_name⚠

Returns the name used to identify the parameter The returned string should be freed using duckdb_free.

duckdb_pending_error⚠

Returns the error message contained within the pending result.

duckdb_pending_execute_check_state⚠

If this returns DUCKDB_PENDING_RESULT_READY, the duckdb_execute_pending function can be called to obtain the result. If this returns DUCKDB_PENDING_RESULT_NOT_READY, the duckdb_pending_execute_check_state function should be called again. If this returns DUCKDB_PENDING_ERROR, an error occurred during execution.

duckdb_pending_execute_task⚠

Executes a single task within the query, returning whether or not the query is ready.

duckdb_pending_execution_is_finished⚠

Returns whether a duckdb_pending_state is finished executing. For example if pending_state is DUCKDB_PENDING_RESULT_READY, this function will return true.

duckdb_pending_prepared⚠

Executes the prepared statement with the given bound parameters, and returns a pending result. The pending result represents an intermediate structure for a query that is not yet fully executed. The pending result can be used to incrementally execute a query, returning control to the client between tasks.

duckdb_pending_prepared_streaming⚠

Executes the prepared statement with the given bound parameters, and returns a pending result. This pending result will create a streaming duckdb_result when executed. The pending result represents an intermediate structure for a query that is not yet fully executed.

duckdb_prepare⚠

Create a prepared statement object from a query.

duckdb_prepare_error⚠

Returns the error message associated with the given prepared statement. If the prepared statement has no error message, this returns nullptr instead.

duckdb_prepare_extracted_statement⚠

Prepare an extracted statement. Note that after calling duckdb_prepare_extracted_statement, the prepared statement should always be destroyed using duckdb_destroy_prepare, even if the prepare fails.

duckdb_prepared_arrow_schema⚠

Fetch the internal arrow schema from the prepared statement. Remember to call release on the respective ArrowSchema object.

duckdb_prepared_statement_type⚠

Returns the statement type of the statement to be executed

duckdb_query⚠

Executes a SQL query within a connection and stores the full (materialized) result in the out_result pointer. If the query fails to execute, DuckDBError is returned and the error message can be retrieved by calling duckdb_result_error.

duckdb_query_arrow⚠

Executes a SQL query within a connection and stores the full (materialized) result in an arrow structure. If the query fails to execute, DuckDBError is returned and the error message can be retrieved by calling duckdb_query_arrow_error.

duckdb_query_arrow_array⚠

Fetch an internal arrow struct array from the arrow result. Remember to call release on the respective ArrowArray object.

duckdb_query_arrow_error⚠

Returns the error message contained within the result. The error is only set if duckdb_query_arrow returns DuckDBError.

duckdb_query_arrow_schema⚠

Fetch the internal arrow schema from the arrow result. Remember to call release on the respective ArrowSchema object.

duckdb_query_progress⚠

Get progress of the running query

duckdb_register_table_function⚠

Register the table function object within the given connection.

duckdb_replacement_scan_add_parameter⚠

Adds a parameter to the replacement scan function.

duckdb_replacement_scan_set_error⚠

Report that an error has occurred while executing the replacement scan.

duckdb_replacement_scan_set_function_name⚠

Sets the replacement function name. If this function is called in the replacement callback, the replacement scan is performed. If it is not called, the replacement callback is not performed.

duckdb_result_arrow_array⚠

Convert a data chunk into an arrow struct array. Remember to call release on the respective ArrowArray object.

duckdb_result_chunk_count⚠

Returns the number of data chunks present in the result.

duckdb_result_error⚠

Returns the error message contained within the result. The error is only set if duckdb_query returns DuckDBError.

duckdb_result_get_chunk⚠

Fetches a data chunk from the duckdb_result. This function should be called repeatedly until the result is exhausted.

duckdb_result_is_streaming⚠

Checks if the type of the internal result is StreamQueryResult.

duckdb_result_return_type⚠

Returns the return_type of the given result, or DUCKDB_RETURN_TYPE_INVALID on error

duckdb_result_statement_type⚠

Returns the statement type of the statement that was executed

duckdb_row_count⚠

Returns the number of rows present in the result object.

duckdb_rows_changed⚠

Returns the number of rows changed by the query stored in the result. This is relevant only for INSERT/UPDATE/DELETE queries. For other queries the rows_changed will be 0.

duckdb_set_config⚠

Sets the specified option for the specified configuration. The configuration option is indicated by name. To obtain a list of config options, see duckdb_get_config_flag.

duckdb_stream_fetch_chunk⚠

Fetches a data chunk from the (streaming) duckdb_result. This function should be called repeatedly until the result is exhausted.

duckdb_string_is_inlined⚠

Whether or not the duckdb_string_t value is inlined. This means that the data of the string does not have a separate allocation.

duckdb_struct_type_child_count⚠

Returns the number of children of a struct type.

duckdb_struct_type_child_name⚠

Retrieves the name of the struct child.

duckdb_struct_type_child_type⚠

Retrieves the child type of the given struct type at the specified index.

duckdb_struct_vector_get_child⚠

Retrieves the child vector of a struct vector.

duckdb_table_function_add_named_parameter⚠

Adds a named parameter to the table function.

duckdb_table_function_add_parameter⚠

Adds a parameter to the table function.

duckdb_table_function_set_bind⚠

Sets the bind function of the table function.

duckdb_table_function_set_extra_info⚠

Assigns extra information to the table function that can be fetched during binding, etc.

duckdb_table_function_set_function⚠

Sets the main function of the table function.

duckdb_table_function_set_init⚠

Sets the init function of the table function.

duckdb_table_function_set_local_init⚠

Sets the thread-local init function of the table function.

duckdb_table_function_set_name⚠

Sets the name of the given table function.

duckdb_table_function_supports_projection_pushdown⚠

Sets whether or not the given table function supports projection pushdown.

duckdb_task_state_is_finished⚠

Check if the provided duckdb_task_state has finished execution

duckdb_to_date⚠

Re-compose a duckdb_date from year, month and date (duckdb_date_struct).

duckdb_to_time⚠

Re-compose a duckdb_time from hour, minute, second and microsecond (duckdb_time_struct).

duckdb_to_timestamp⚠

Re-compose a duckdb_timestamp from a duckdb_timestamp_struct.

duckdb_uhugeint_to_double⚠

Converts a duckdb_uhugeint object (as obtained from a DUCKDB_TYPE_UHUGEINT column) into a double.

duckdb_union_type_member_count⚠

Returns the number of members that the union type has.

duckdb_union_type_member_name⚠

Retrieves the name of the union member.

duckdb_union_type_member_type⚠

Retrieves the child type of the given union member at the specified index.

duckdb_validity_row_is_valid⚠

Returns whether or not a row is valid (i.e. not NULL) in the given validity mask.

duckdb_validity_set_row_invalid⚠

In a validity mask, sets a specific row to invalid.

duckdb_validity_set_row_valid⚠

In a validity mask, sets a specific row to valid.

duckdb_validity_set_row_validity⚠

In a validity mask, sets a specific row to either valid or invalid.

duckdb_value_blob⚠

returns: The duckdb_blob value at the specified location. Returns a blob with blob.data set to nullptr if the value cannot be converted. The resulting field “blob.data” must be freed with duckdb_free.

duckdb_value_boolean⚠

returns: The boolean value at the specified location, or false if the value cannot be converted.

duckdb_value_date⚠

returns: The duckdb_date value at the specified location, or 0 if the value cannot be converted.

duckdb_value_decimal⚠

returns: The duckdb_decimal value at the specified location, or 0 if the value cannot be converted.

duckdb_value_double⚠

returns: The double value at the specified location, or 0 if the value cannot be converted.

duckdb_value_float⚠

returns: The float value at the specified location, or 0 if the value cannot be converted.

duckdb_value_hugeint⚠

returns: The duckdb_hugeint value at the specified location, or 0 if the value cannot be converted.

duckdb_value_int8⚠

returns: The int8_t value at the specified location, or 0 if the value cannot be converted.

duckdb_value_int16⚠

returns: The int16_t value at the specified location, or 0 if the value cannot be converted.

duckdb_value_int32⚠

returns: The int32_t value at the specified location, or 0 if the value cannot be converted.

duckdb_value_int64⚠

returns: The int64_t value at the specified location, or 0 if the value cannot be converted.

duckdb_value_interval⚠

returns: The duckdb_interval value at the specified location, or 0 if the value cannot be converted.

duckdb_value_is_null⚠

returns: Returns true if the value at the specified index is NULL, and false otherwise.

duckdb_value_string⚠

returns: The string value at the specified location. The resulting field “string.data” must be freed with duckdb_free.

duckdb_value_string_internal⚠

DEPRECATED: use duckdb_value_string_internal instead. This function does not work correctly if the string contains null bytes. returns: The char* value at the specified location. ONLY works on VARCHAR columns and does not auto-cast. If the column is NOT a VARCHAR column this function will return NULL.

duckdb_value_time⚠

returns: The duckdb_time value at the specified location, or 0 if the value cannot be converted.

duckdb_value_timestamp⚠

returns: The duckdb_timestamp value at the specified location, or 0 if the value cannot be converted.

duckdb_value_uhugeint⚠

returns: The duckdb_uhugeint value at the specified location, or 0 if the value cannot be converted.

duckdb_value_uint8⚠

returns: The uint8_t value at the specified location, or 0 if the value cannot be converted.

duckdb_value_uint16⚠

returns: The uint16_t value at the specified location, or 0 if the value cannot be converted.

duckdb_value_uint32⚠

returns: The uint32_t value at the specified location, or 0 if the value cannot be converted.

duckdb_value_uint64⚠

returns: The uint64_t value at the specified location, or 0 if the value cannot be converted.

duckdb_value_varchar⚠

DEPRECATED: use duckdb_value_string instead. This function does not work correctly if the string contains null bytes. returns: The text value at the specified location as a null-terminated string, or nullptr if the value cannot be converted. The result must be freed with duckdb_free.

duckdb_value_varchar_internal⚠

DEPRECATED: use duckdb_value_string_internal instead. This function does not work correctly if the string contains null bytes. returns: The char* value at the specified location. ONLY works on VARCHAR columns and does not auto-cast. If the column is NOT a VARCHAR column this function will return NULL.

duckdb_vector_assign_string_element⚠

Assigns a string element in the vector at the specified location.

duckdb_vector_assign_string_element_len⚠

Assigns a string element in the vector at the specified location. You may also use this function to assign BLOBs.

duckdb_vector_ensure_validity_writable⚠

Ensures the validity mask is writable by allocating it.

duckdb_vector_get_column_type⚠

Retrieves the column type of the specified vector.

duckdb_vector_get_data⚠

Retrieves the data pointer of the vector.

duckdb_vector_get_validity⚠

Retrieves the validity mask pointer of the specified vector.

duckdb_vector_size⚠

The internal vector size used by DuckDB. This is the amount of tuples that will fit into a data chunk created by duckdb_create_data_chunk.

Type Aliases

DUCKDB_TYPE

! An enum over DuckDB’s internal types.

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

duckdb_appender

! The appender enables fast data loading into DuckDB. ! Must be destroyed with duckdb_appender_destroy.

duckdb_arrow

! Holds an arrow query result. Must be destroyed with duckdb_destroy_arrow.

duckdb_arrow_array

! Holds an arrow array. Remember to release the respective ArrowArray object.

duckdb_arrow_schema

! Holds an arrow schema. Remember to release the respective ArrowSchema object.

duckdb_arrow_stream

! Holds an arrow array stream. Must be destroyed with duckdb_destroy_arrow_stream.

duckdb_bind_info

! The bind info of the function. When setting this info, it is necessary to pass a destroy-callback function.

duckdb_config

! Can be used to provide start-up options for the DuckDB instance. ! Must be destroyed with duckdb_destroy_config.

duckdb_connection

! A connection to a duckdb database. Must be closed with duckdb_disconnect.

duckdb_data_chunk

! Contains a data chunk from a duckdb_result. ! Must be destroyed with duckdb_destroy_data_chunk.

duckdb_database

! A database object. Should be closed with duckdb_close.

duckdb_delete_callback_t

! The callback that will be called to destroy data, e.g., ! bind data (if any), init data (if any), extra data for replacement scans (if any)

duckdb_extracted_statements

! Extracted statements. Must be destroyed with duckdb_destroy_extracted.

duckdb_function_info

! Additional function info. When setting this info, it is necessary to pass a destroy-callback function.

duckdb_init_info

! Additional function init info. When setting this info, it is necessary to pass a destroy-callback function.

duckdb_logical_type

! Holds an internal logical type. ! Must be destroyed with duckdb_destroy_logical_type.

duckdb_pending_result

! The pending result represents an intermediate structure for a query that is not yet fully executed. ! Must be destroyed with duckdb_destroy_pending.

duckdb_pending_state

! An enum over the pending state of a pending query result.

duckdb_prepared_statement

! A prepared statement is a parameterized query that allows you to bind parameters to it. ! Must be destroyed with duckdb_destroy_prepare.

duckdb_replacement_callback_t

! A replacement scan function that can be added to a database.

duckdb_replacement_scan_info

! Additional replacement scan info. When setting this info, it is necessary to pass a destroy-callback function.

duckdb_result_type

! An enum over DuckDB’s different result types.

duckdb_state

! An enum over the returned state of different functions.

duckdb_statement_type

! An enum over DuckDB’s different statement types.

duckdb_table_function

! A table function. Must be destroyed with duckdb_destroy_table_function.

duckdb_table_function_bind_t

! The bind function of the table function.

duckdb_table_function_init_t

! The (possibly thread-local) init function of the table function.

duckdb_table_function_t

! The main function of the table function.

duckdb_task_state

! Used for threading, contains a task state. Must be destroyed with duckdb_destroy_state.

duckdb_type

! An enum over DuckDB’s internal types. ! An enum over DuckDB’s internal types.

duckdb_value

! Holds a DuckDB value, which wraps a type. ! Must be destroyed with duckdb_destroy_value.

duckdb_vector

! A vector to a specified column in a data chunk. Lives as long as the ! data chunk lives, i.e., must not be destroyed.

idx_t

! DuckDB’s index type.

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t

wchar_t

Unions

duckdb_string_t__bindgen_ty_1